+++ test/unit/client-bridge.test.js
@@ -1,11 +1,11 @@
 'use strict';
 var q = require('q'),
-    ClientBridge = require('../lib/browser/client-bridge'),
-    StateError = require('../lib/errors/state-error'),
+    ClientBridge = require('../../lib/browser/client-bridge'),
+    StateError = require('../../lib/errors/state-error'),
     assert = require('chai').assert,
     sinon = require('sinon'),
 
-    makeBrowser = require('./util').makeBrowser,
+    makeBrowser = require('../util').makeBrowser,
     CALL = '__gemini.example(1, "two")';
 
 describe('ClientBridge', function() {
