+++ test/unit/config-options.test.js
@@ -1,8 +1,8 @@
 'use strict';
-var Config = require('../lib/config'),
-    parser = require('../lib/config/options'),
+var Config = require('../../lib/config'),
+    parser = require('../../lib/config/options'),
     assert = require('chai').assert,
-    GeminiError = require('../lib/errors/gemini-error'),
+    GeminiError = require('../../lib/errors/gemini-error'),
     MissingOptionError = require('gemini-configparser').MissingOptionError,
     sinon = require('sinon'),
     _ = require('lodash');
