+++ test/unit/stats.test.js
@@ -5,8 +5,8 @@ var EventEmitter = require('events').EventEmitter,
 
     inherit = require('inherit'),
 
-    Stats = require('../lib/stats'),
-    RunnerEvents = require('../lib/constants/runner-events');
+    Stats = require('../../lib/stats'),
+    RunnerEvents = require('../../lib/constants/runner-events');
 
 describe('stats', function() {
     var stats;
